test(chore): start of converting mcp configure tests
